+++ lib/view_component/base.rb
@@ -23,6 +23,13 @@ module ViewComponent
     class_attribute :content_areas
     self.content_areas = [] # class_attribute:default doesn't work until Rails 5.2
 
+    # Hook for allowing components to do work as part of the compilation process.
+    #
+    # For example, one might compile component-specific assets at this point.
+    def self.after_compile
+      # noop
+    end
+
     # Entrypoint for rendering components.
     #
     # view_context: ActionView context from calling view